formacao_java_e_orientacao_a_objetos/java_tdd/codigo/java-testes-unitarios-tdd-projeto_refatorando_test/src/test/java/br/com/alura/tdd/service/ReajusteServiceTest.java
package br.com.alura.tdd.service;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.math.BigDecimal;
import java.time.LocalDate;

import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import br.com.alura.tdd.enums.Desempenho;
import br.com.alura.tdd.modelo.Funcionario;

public class ReajusteServiceTest {

	private ReajusteService service;
	private Funcionario funcionario;
	
	@BeforeAll
	public static void inicio() {
		System.out.println("Inicio");
	}
	
	@BeforeEach
	public void inicializar() {
		System.out.println("Inicializando...");
		this.service = new ReajusteService();
		this.funcionario = new Funcionario("Fernando", LocalDate.now(), new BigDecimal("100.00"));
	}

	@Test
	public void reajusteDeveriaSerDeTresPorcentoQuandoDesempenhoForADesejar() {
		service.concederReajuste(funcionario, Desempenho.A_DESEJAR);
		assertEquals(new BigDecimal("103.00"), funcionario .getSalario());
	}
	
	@Test
	public void reajusteDeveriaSerDeQuinzePorcentoQuandoDesempenhoForBom() {
		service.concederReajuste(funcionario, Desempenho.BOM);
		assertEquals(new BigDecimal("115.00"), funcionario .getSalario());
	}
	
	@Test
	public void reajusteDeveriaSerDeVintePorcentoQuandoDesempenhoForOtimo() {
		service.concederReajuste(funcionario, Desempenho.OTIMO);
		assertEquals(new BigDecimal("120.00"), funcionario .getSalario());
	}
	
	@AfterEach
	public void finalizando() {
		System.out.println("Finalizando...");
	}
	
	@AfterAll
	public static void termino() {
		System.out.println("Termino");
	}
	
}
